+++ src/resources/receipts.ts
@@ -5,9 +5,6 @@ import {
   ReceiptsToInvoiceInput,
   Receipt,
   SearchResult,
-  CursorSearchResult,
-  CursorSearchParams,
-  PageSearchParams,
   SendEmailBody,
   PreviewReceiptsToInvoicePdfInput,
 } from '../types'
@@ -33,12 +30,7 @@ export default class Receipts {
    * @param params Search parameters
    * @returns Search results object. The object contains a `data` property with the list of receipts.
    */
-  list(params: CursorSearchParams): Promise<CursorSearchResult<Receipt>>;
-  list(params: PageSearchParams): Promise<SearchResult<Receipt>>;
-  list(params?: Record<string, any> | null): Promise<SearchResult<Receipt>>;
-  list(
-    params?: Record<string, any> | null,
-  ): Promise<SearchResult<Receipt> | CursorSearchResult<Receipt>> {
+  list(params?: Record<string, any> | null): Promise<SearchResult<Receipt>> {
     if (!params) params = {}
     return this.client.get('/receipts', { params })
   }